fix(cybervision): fix segmentation WGSL type errors and add test coverage
Fix two bugs in segmentation effect:
1. WGSL type error: vec2<f32> constructed from u32 params without f32() cast
2. Revert maskTexture format from r8unorm to rgba8unorm for compatibility

Add missing test coverage for segmentation effect:
- Add segmentation.wgsl to unit test shader list
- Add renderSegmentation to integration test required methods
- Add segmentation to E2E test helpers and effect lists

+++ apps/cybervision/webroot/static/shaders/segmentation.wgsl
@@ -56,7 +56,7 @@ fn sampleMaskSoft(pos: vec2<i32>) -> f32 {
   if (params.softEdges == 0u) {
     // Standard nearest-neighbor sampling
     let maskPos = vec2<i32>(
-      vec2<f32>(pos) * 256.0 / vec2<f32>(params.width, params.height)
+      vec2<f32>(pos) * 256.0 / vec2<f32>(f32(params.width), f32(params.height))
     );
     let maskColor = textureLoad(maskTexture, maskPos, 0);
     return maskColor.r;
@@ -91,7 +91,7 @@ fn detectEdge(pos: vec2<i32>, maskValue: f32) -> f32 {
 
   // Sample mask at multiple offsets to detect edges
   let maskPos = vec2<i32>(
-    vec2<f32>(pos) * 256.0 / vec2<f32>(params.width, params.height)
+    vec2<f32>(pos) * 256.0 / vec2<f32>(f32(params.width), f32(params.height))
   );
 
   var edgeStrength = 0.0;
